Feinschliff - Notizbuch löschen logik implementierung und done

+++ backend/routes/notebookRoutes.js
@@ -3,6 +3,9 @@ const router = express.Router();
 
 // import Notebook schema
 const Notebook = require("../models/noteBookModel");
+// import Note schema for safe delete
+const Note = require("../models/noteModel");
+
 
 // get all notebooks
 router.get("/", async (req, res) => {
@@ -23,11 +26,15 @@ router.get("/:id", async (req, res) => {
 
 // create one
 router.post("/", async (req, res) => {
-  const notebook = await Notebook.create({
-    title: req.body.title,
-    description: req.body.description,
-  });
-  res.status(201).json(notebook);
+  try {
+    const notebook = await Notebook.create({
+      title: req.body.title,
+      description: req.body.description,
+    });
+    res.status(201).json(notebook);
+  } catch (err) {
+    res.status(405).send();
+  }
 });
 
 // update one
@@ -40,17 +47,21 @@ router.put("/:id", async (req, res) => {
   }
 
   // if the notebook is available, then update it
-  const updatedNotebook = await Notebook.findByIdAndUpdate(
-    req.params.id,
-    req.body,
-    {
-      title: req.body.title,
-      description: req.body.description,
-    }
-  );
+  try {
+    const updatedNotebook = await Notebook.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      {
+        title: req.body.title,
+        description: req.body.description,
+      }
+    );
 
-  // return the updated notebook
-  res.json(updatedNotebook);
+    // return the updated notebook
+    res.json(updatedNotebook);
+  } catch (err) {
+    res.status(405).send();
+  }
 });
 
 // delete one
@@ -62,10 +73,14 @@ router.delete("/:id", async (req, res) => {
     // if it is not available, return 404 - Not Found
     res.status(404).send();
   }
-
-  // if it is available, then delete it
-  let notebook = await Notebook.findByIdAndDelete(req.params.id);
-  res.json(notebook);
+  let associatedNotes = await Note.find({ notebookId: req.params.id })
+  if(associatedNotes.length === 0){
+    // if it is available, then delete it
+    let notebook = await Notebook.findByIdAndDelete(req.params.id);
+    res.json(notebook);
+  }else{
+    res.status(405).send();
+  }
 });
 
 module.exports = router;
